+++ src/main/kotlin/lotto/LottoApp.kt
@@ -1,5 +1,12 @@
 package lotto
 
+import lotto.domain.LottoMachine
+import lotto.domain.LottoStatistics
+import lotto.domain.Winning
+import lotto.domain.WinningMatcher
+import lotto.view.InputView
+import lotto.view.ResultView
+
 fun main() {
     val money = InputView().getMoney()
     val tickets = LottoMachine().buy(money)
