+++ src/main/kotlin/lotto/domain/LottoStatistics.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import kotlin.math.round
