+++ src/test/kotlin/lotto/LottoMachineTest.kt
@@ -3,6 +3,7 @@ package lotto
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
+import lotto.domain.LottoMachine
 import java.lang.IllegalArgumentException
 
 class LottoMachineTest : FreeSpec({
